Add extension methods to help find associated tileset for given tile.
